Admin management

+++ TAABP.API/Controllers/CreditCardController.cs
@@ -9,7 +9,7 @@ using ILogger = Serilog.ILogger;
 
 namespace TAABP.API.Controllers
 {
-    [Route("api/{userId}/CreditCards")]
+    [Route("api/Users/{userId}/CreditCards")]
     [ApiController]
     [Authorize]
     public class CreditCardController : ControllerBase
@@ -17,19 +17,28 @@ namespace TAABP.API.Controllers
         private readonly ICreditCardService _creditCardService;
         private readonly ILogger _logger;
         private readonly IValidator<CreditCardDto> _creditCardValidator;
-        public CreditCardController(ICreditCardService creditCardService, IValidator<CreditCardDto> validator)
+        private readonly IUserService _userService;
+        public CreditCardController(ICreditCardService creditCardService,
+            IValidator<CreditCardDto> validator,
+            IUserService userService)
         {
             _creditCardService = creditCardService;
             _logger = Log.ForContext<CreditCardController>();
             _creditCardValidator = validator;
+            _userService = userService;
         }
 
         [HttpGet("{paymentOptionId}")]
-        public async Task<IActionResult> GetPaymentOptionByIdAsync(int paymentOptionId)
+        public async Task<IActionResult> GetPaymentOptionByIdAsync(string userId, int paymentOptionId)
         {
             _logger.Information("Fetching payment option with ID {PaymentOptionId}", paymentOptionId);
             try
             {
+                if(userId != _userService.GetCurrentUserId())
+                {
+                    _logger.Warning("Unauthorized access to payment option with ID {PaymentOptionId}", paymentOptionId);
+                    return Unauthorized();
+                }
                 var paymentOption = await _creditCardService.GetPaymentOptionByIdAsync(paymentOptionId);
                 _logger.Information("Successfully fetched payment option with ID {PaymentOptionId}", paymentOptionId);
                 return Ok(paymentOption);
@@ -52,6 +61,11 @@ namespace TAABP.API.Controllers
             _logger.Information("Adding new payment option for user with ID {UserId}", userId);
             try
             {
+                if (userId != _userService.GetCurrentUserId())
+                {
+                    _logger.Warning("Unauthorized access to add payment option for user with ID {UserId}", userId);
+                    return Unauthorized();
+                }
                 await _creditCardValidator.ValidateAndThrowAsync(paymentOption);
                 int cardId = await _creditCardService.AddNewPaymentOptionAsync(userId, paymentOption);
                 var card = await _creditCardService.GetPaymentOptionByIdAsync(cardId);
@@ -76,6 +90,11 @@ namespace TAABP.API.Controllers
             _logger.Information("Updating payment option with ID {PaymentOptionId} for user with ID {UserId}", creditcardId, userId);
             try
             {
+                if (userId != _userService.GetCurrentUserId())
+                {
+                    _logger.Warning("Unauthorized access to update payment option with ID {PaymentOptionId} for user with ID {UserId}", creditcardId, userId);
+                    return Unauthorized();
+                }
                 await _creditCardValidator.ValidateAndThrowAsync(paymentOption);
                 await _creditCardService.UpdatePaymentOptionAsync(creditcardId, userId, paymentOption);
                 _logger.Information("Successfully updated payment option with ID {PaymentOptionId} for user with ID {UserId}", creditcardId, userId);
@@ -99,6 +118,11 @@ namespace TAABP.API.Controllers
             _logger.Information("Deleting payment option with ID {PaymentOptionId} for user with ID {UserId}", creditCardId, userId);
             try
             {
+                if (userId != _userService.GetCurrentUserId())
+                {
+                    _logger.Warning("Unauthorized access to delete payment option with ID {PaymentOptionId} for user with ID {UserId}", creditCardId, userId);
+                    return Unauthorized();
+                }
                 await _creditCardService.DeletePaymentOptionAsync(userId, creditCardId);
                 _logger.Information("Successfully deleted payment option with ID {PaymentOptionId} for user with ID {UserId}", creditCardId, userId);
                 return NoContent();
